Function refactoring and reordering
