--- resources/lib/bitrate_dialog.py
@@ -1,57 +1,0 @@
-from __future__ import division, absolute_import, print_function, unicode_literals
-
-import xbmc
-import xbmcgui
-
-from .loghandler import LazyLogger
-
-log = LazyLogger(__name__)
-
-
-class BitrateDialog(xbmcgui.WindowXMLDialog):
-
-    slider_control = None
-    bitrate_label = None
-    initial_bitrate_value = 0
-    selected_transcode_value = 0
-
-    def __init__(self, *args, **kwargs):
-        log.debug("BitrateDialog: __init__")
-        xbmcgui.WindowXML.__init__(self, *args, **kwargs)
-
-    def onInit(self):
-        log.debug("ActionMenu: onInit")
-        self.action_exitkeys_id = [10, 13]
-
-        self.slider_control = self.getControl(3000)
-        self.slider_control.setInt(self.initial_bitrate_value, 400, 100, 15000)
-
-        self.bitrate_label = self.getControl(3030)
-        bitrate_label_string = str(self.slider_control.getInt()) + " Kbs"
-        self.bitrate_label.setLabel(bitrate_label_string)
-
-    def onFocus(self, control_id):
-        pass
-
-    def doAction(self, action_id):
-        pass
-
-    def onMessage(self, message):
-        log.debug("ActionMenu: onMessage: {0}".format(message))
-
-    def onAction(self, action):
-
-        bitrate_label_string = str(self.slider_control.getInt()) + " Kbs"
-        self.bitrate_label.setLabel(bitrate_label_string)
-
-        if action.getId() == 10:  # ACTION_PREVIOUS_MENU
-            self.close()
-        elif action.getId() == 92:  # ACTION_NAV_BACK
-            self.close()
-        elif action.getId() == 7:  # ENTER
-            self.selected_transcode_value = self.slider_control.getInt()
-            self.close()
-
-    def onClick(self, control_id):
-        if control_id == 3000:
-            log.debug("ActionMenu: Selected Item: {0}".format(control_id))
