+++ service.py
@@ -11,11 +11,10 @@ import xbmcgui
 from resources.lib.downloadutils import DownloadUtils, save_user_details
 from resources.lib.loghandler import LazyLogger
 from resources.lib.play_utils import Service, PlaybackService, send_progress
-from resources.lib.kodi_utils import HomeWindow
+from resources.lib.kodi_utils import HomeWindow, ContextMonitor
 from resources.lib.widgets import set_background_image, set_random_movies
 from resources.lib.websocket_client import WebSocketClient
 from resources.lib.menu_functions import set_library_window_values
-from resources.lib.context_monitor import ContextMonitor
 from resources.lib.server_detect import check_server, check_safe_delete_available, check_connection_speed
 from resources.lib.library_change_monitor import LibraryChangeMonitor
 from resources.lib.datamanager import clear_old_cache_data
